+++ tests/run.lua
@@ -11,6 +11,7 @@ package.path = here .. "/?.lua;" .. package.path
 local lu = require("luaunit")
 require("test_add_code_files")
 require("test_crossref_anx")
+require("test_docx_cell_paragraph")
 require("test_filetree")
 require("test_r_syntax")